Update cli to use gflags and algorithm choice
* Converts the existing CLI system to use the gflags module rather
than using argc/argv directly

* Now prints a slightly more useful usage message when neecesary
cli flags are missing

* Cleans up the main by moving parsing files into their own header
file and has them return map rather than using a single global map
(note we should probably update this to pass a map reference for
better performance in the future)
