+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "googletest"]
 	path = googletest
 	url = https://github.com/google/googletest.git
+[submodule "gflags"]
+	path = gflags
+	url = https://github.com/gflags/gflags.git
